+++ cmd/listings.go
@@ -23,52 +23,325 @@ THE SOFTWARE.
 package cmd
 
 import (
-	"fmt"
+	"errors"
+	"time"
 
+	"github.com/jedib0t/go-pretty/v6/table"
+	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+
+	"github.com/asphaltbuffet/ogma/pkg/datastore"
 )
 
-var (
-	searchYear   int
-	searchIssue  int
-	searchMember int
-)
+// A Listing contains relevant information for LEX listings.
+type Listing struct {
+	ID                  int `storm:"id,increment"`
+	Volume              int
+	IssueNumber         int
+	Year                int
+	PageNumber          int
+	IndexedCategory     string `storm:"index"`
+	IndexedMemberNumber int    `storm:"index"`
+	MemberExtension     string
+	IsInternational     bool
+	IsReview            bool
+	ListingText         string
+	IsArt               bool
+	IsFlagged           bool
+}
 
 // listingsCmd represents the listings command.
 var listingsCmd = &cobra.Command{
 	Use:   "listings",
-	Short: "Search listings by specified field.",
-	Long: `Returns all listing information based on search criteria.
-	
-	Available search criteria:
-	  Year
-	  LEX Issue
-	  Member Number
-	  Category - NOT IMPLEMENTED
-	  `,
-	Run: func(cmd *cobra.Command, args []string) {
-		if searchYear >= viper.GetInt("search.min_year") {
-			fmt.Println("Searching listings by year: ", searchYear)
-		} else {
-			fmt.Println("Searching listings by year: ANY")
-		}
+	Short: "Access listings functionality.",
+	Long:  ``,
+}
 
-		if searchIssue >= viper.GetInt("search.min_issue") {
-			fmt.Println("Searching listings by issue: ", searchIssue)
-		} else {
-			fmt.Println("Searching listings by issue: ANY")
-		}
-
-		if searchMember >= viper.GetInt("search.min_member") {
-			fmt.Println("Searching listings by member: ", searchMember)
-		} else {
-			fmt.Println("Searching listings by member: ANY")
-		}
+var searchListingsCmd = &cobra.Command{
+	Use:   "search",
+	Short: "Returns all listing information based on search criteria.",
+	Long:  ``,
+	Args:  cobra.NoArgs,
+	RunE: func(cmd *cobra.Command, args []string) error {
+		return RunSearchListings(cmd)
 	},
 }
 
+var addListingCmd = &cobra.Command{
+	Use:   "add",
+	Short: "Add a single listing.",
+	Long:  ``,
+	Args:  cobra.NoArgs,
+	RunE: func(cmd *cobra.Command, args []string) error {
+		return RunAddListing(cmd)
+	},
+}
+
+// RunSearchListings performs action associated with listings application command.
+func RunSearchListings(cmd *cobra.Command) error {
+	year, err := cmd.Flags().GetInt("year")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "year",
+		}).Error("Invalid flag.")
+		return err
+	}
+
+	issue, err := cmd.Flags().GetInt("issue")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "issue",
+		}).Error("Invalid flag.")
+		return err
+	}
+
+	member, err := cmd.Flags().GetInt("member")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "member",
+		}).Error("Invalid flag.")
+		return err
+	}
+
+	log.WithFields(log.Fields{
+		"cmd":    "listings.search",
+		"year":   year,
+		"issue":  issue,
+		"member": member,
+	}).Info("Searching listings.")
+
+	// TODO: SearchListings should return an error too.
+	resultsCount := SearchListings(year, issue, member)
+	cmd.Println("Found", resultsCount, "listings.")
+
+	if resultsCount > viper.GetInt("search.max_results") {
+		log.WithFields(log.Fields{
+			"cmd":          "listings.search",
+			"resultsCount": resultsCount,
+			"maxResults":   viper.GetInt("search.max_results"),
+		}).Warn("Query return too large.")
+		return errors.New("too many results")
+	}
+
+	return nil
+}
+
+// RunAddListing adds a single listing to the datastore.
+func RunAddListing(cmd *cobra.Command) error { //nolint:funlen // TODO: refactor later 2021-10-31 BL
+	dsManager, err := datastore.New(viper.GetString("datastore.filename"))
+	if err != nil {
+		log.Fatal("Datastore manager failure.")
+	}
+
+	defer dsManager.Stop()
+	if err != nil {
+		log.Error("Failed to save to db: ")
+	}
+
+	volume, err := cmd.Flags().GetInt("volume")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "volume",
+		}).Error("Invalid flag.")
+		return err
+	}
+
+	lex, err := cmd.Flags().GetInt("lex")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "lex",
+		}).Error("Invalid flag.")
+		return err
+	}
+
+	year, err := cmd.Flags().GetInt("year")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "year",
+		}).Error("Invalid flag.")
+		return err
+	}
+
+	page, err := cmd.Flags().GetInt("page")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "page",
+		}).Error("Invalid flag.")
+		return err
+	}
+	category, err := cmd.Flags().GetString("category")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "category",
+		}).Error("Invalid flag.")
+		return err
+	}
+
+	member, err := cmd.Flags().GetInt("member")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "member",
+		}).Error("Invalid flag.")
+		return err
+	}
+
+	international, err := cmd.Flags().GetBool("international")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "international",
+		}).Error("Invalid flag.")
+		return err
+	}
+
+	review, err := cmd.Flags().GetBool("review")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "review",
+		}).Error("Invalid flag.")
+		return err
+	}
+
+	text, err := cmd.Flags().GetString("text")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "year",
+		}).Error("Invalid flag.")
+		return err
+	}
+
+	sketch, err := cmd.Flags().GetBool("sketch")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "sketch",
+		}).Error("Invalid flag.")
+		return err
+	}
+
+	flag, err := cmd.Flags().GetBool("flag")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"flag": "flag",
+		}).Error("Invalid flag.")
+		return err
+	}
+
+	log.WithFields(log.Fields{
+		"cmd":    "listings.add",
+		"year":   year,
+		"issue":  lex,
+		"member": member,
+	}).Info("Adding a listing.")
+
+	newListing := Listing{
+		Volume:              volume,
+		IssueNumber:         lex,
+		Year:                year,
+		PageNumber:          page,
+		IndexedCategory:     category,
+		IndexedMemberNumber: member,
+		MemberExtension:     "", // TODO: add support for member extensions
+		IsInternational:     international,
+		IsReview:            review,
+		ListingText:         text,
+		IsArt:               sketch,
+		IsFlagged:           flag,
+	}
+
+	err = dsManager.Store.Save(&newListing)
+	if err != nil {
+		log.Error("Failed to save new listing.")
+	}
+	// TODO: add formatted listing as output.
+	cmd.Println("Added a listing.")
+
+	lt := table.NewWriter()
+	lt.AppendHeader(table.Row{
+		"Volume",
+		"Issue",
+		"Year",
+		"Page",
+		"Category",
+		"Member",
+		"International",
+		"Review",
+		"Text",
+		"Sketch",
+		"Flagged",
+	})
+	lt.AppendRow([]interface{}{
+		newListing.Volume,
+		newListing.IssueNumber,
+		newListing.Year,
+		newListing.PageNumber,
+		newListing.IndexedCategory,
+		newListing.IndexedMemberNumber,
+		newListing.IsInternational,
+		newListing.IsReview,
+		newListing.ListingText,
+		newListing.IsArt,
+		newListing.IsFlagged,
+	})
+	lt.SetColumnConfigs([]table.ColumnConfig{
+		{
+			Name:     "Text",
+			WidthMax: viper.GetInt("defaults.max_column"),
+		},
+	})
+
+	// TODO: wrap output styles in a new flag
+	// lt.SetStyle(table.StyleColoredBright)
+	cmd.Println(lt.Render())
+
+	// Actually return error if add was successful.
+	return nil
+}
+
+// SearchListings queries listing db for matching listings.
+// TODO: Actually interact with db so it's functional.
+func SearchListings(y int, i int, m int) (count int) {
+	count = 0
+	if y == 2021 { //nolint:gomnd // placeholder before db query is implemented
+		count += 3
+	} else {
+		count += 0
+	}
+
+	if i == 56 { //nolint:gomnd // placeholder before db query is implemented
+		count += 7
+	} else {
+		count += 0
+	}
+
+	if m == 1000 { //nolint:gomnd // placeholder before db query is implemented
+		count += 2
+	} else {
+		count += 0
+	}
+
+	return count
+}
+
 func init() {
+	addListingCmd.Flags().IntP("volume", "v", -1, "Volume containing listing entry.")
+	addListingCmd.Flags().IntP("lex", "l", viper.GetInt("defaults.issue"), "LEX issue containing listing entry.")
+	addListingCmd.Flags().IntP("year", "y", time.Now().Year(), "Year of listing entry..")
+	// addListingCmd.MarkFlagRequired("year") //nolint:errcheck,gosec // handled by cobra
+	addListingCmd.Flags().IntP("page", "p", -1, "Page number of listing entry.")
+	addListingCmd.Flags().StringP("category", "c", "", "Category of listing entry.")
+	addListingCmd.Flags().IntP("member", "m", -1, "Member number of listing entry.")
+	addListingCmd.Flags().BoolP("international", "i", false, "Is international postage required?")
+	addListingCmd.Flags().BoolP("review", "r", false, "Is this a book review listing entry?")
+	addListingCmd.Flags().StringP("text", "t", "", "Text of listing entry.")
+	addListingCmd.Flags().BoolP("sketch", "s", false, "Is this a sketch listing entry?")
+	addListingCmd.Flags().BoolP("flag", "f", false, "Has this listing entry been flagged?")
+	listingsCmd.AddCommand(addListingCmd)
+
+	searchListingsCmd.Flags().IntP("year", "y", -1, "Search listings by LEX Issue year.")
+	searchListingsCmd.Flags().IntP("issue", "i", -1, "Search listings by LEX Issue Number.")
+	searchListingsCmd.Flags().IntP("member", "m", -1, "Search listings by member number.")
+
+	listingsCmd.AddCommand(searchListingsCmd)
 	rootCmd.AddCommand(listingsCmd)
 
 	// Here you will define your flags and configuration settings.
@@ -80,7 +353,4 @@ func init() {
 	// Cobra supports local flags which will only run when this command
 	// is called directly, e.g.:
 	// listingsCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-	listingsCmd.Flags().IntVarP(&searchYear, "year", "y", viper.GetInt("search.default_year"), "Search listings by LEX Issue year. Use '0' to search 'ANY'")
-	listingsCmd.Flags().IntVarP(&searchIssue, "issue", "i", viper.GetInt("search.default_issue"), "Search listings by LEX Issue Number. Use '0' to search 'ANY'")
-	listingsCmd.Flags().IntVarP(&searchMember, "member", "m", viper.GetInt("search.default_member"), "Search listings by member number. Use '0' to search 'ANY'")
 }