+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.2]
 ### Added
 - Application configuration file (.ogma)
 - Logging
@@ -28,5 +30,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixes
 ### Security
 
-[Unreleased]: https://github.com/asphaltbuffet/ogma/compare/v0.0.1...HEAD
+[Unreleased]: https://github.com/asphaltbuffet/ogma/compare/v0.0.2...HEAD
+[0.0.2]: https://github.com/asphaltbuffet/ogma/releases/tag/v0.0.2
 [0.0.1]: https://github.com/asphaltbuffet/ogma/releases/tag/v0.0.1